techassist-ai/backend/app/services/vision_service.py
import base64
import io
from PIL import Image
import numpy as np
from app.models.schemas import VisionAnalysis

class VisionService:
    def __init__(self):
        pass
    
    async def analyze_image(self, image_base64: str) -> VisionAnalysis:
        try:
            # Decode base64 image
            image_data = base64.b64decode(image_base64.split(',')[1])
            image = Image.open(io.BytesIO(image_data))
            
            # For demo purposes, return mock analysis
            # In production, this would use actual computer vision models
            return VisionAnalysis(
                equipment_type="HVAC Unit - Carrier Model",
                issues_detected=["Possible capacitor bulging", "Rust on exterior panel"],
                confidence=0.85,
                suggestions=[
                    "Inspect the capacitor for physical damage",
                    "Check electrical connections",
                    "Test capacitor with multimeter"
                ]
            )
            
        except Exception as e:
            return VisionAnalysis(
                equipment_type=None,
                issues_detected=None,
                confidence=0.0,
                suggestions=["Unable to analyze image. Please try again with a clearer photo."]
            )