techassist-ai/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.6.1
langchain==0.0.340
langgraph==0.0.20
redis==5.0.1
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
python-socketio==5.10.0
python-engineio==4.8.0
aiofiles==23.2.1
httpx==0.25.1